+++ src/test/java/dinkplugin/notifiers/QuestNotifierTest.java
@@ -13,8 +13,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 
+import static dinkplugin.util.QuestUtils.parseQuestWidget;
 import static net.runelite.api.widgets.WidgetID.QUEST_COMPLETED_GROUP_ID;
 import static net.runelite.api.widgets.WidgetInfo.QUEST_COMPLETED_NAME_TEXT;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.Mockito.mock;
@@ -97,6 +100,26 @@ class QuestNotifierTest extends MockedNotifierTest {
 
         // verify no message
         verify(messageHandler, never()).createMessage(any(), anyBoolean(), any());
+    }
+
+    @Test
+    void parseWidgets() {
+        assertEquals("Recipe for Disaster - Another Cook's Quest", parseQuestWidget("You have completed Another Cook's Quest!"));
+        assertEquals("Recipe for Disaster - Another Cook's Quest", parseQuestWidget("You have assisted the Lumbridge Cook... again!"));
+        assertEquals("Recipe for Disaster - Goblin Generals", parseQuestWidget("You have freed the Goblin Generals!"));
+        assertEquals("Recipe for Disaster - Sir Amik Varze", parseQuestWidget("You have freed Sir Amik Varze!"));
+        assertEquals("Recipe for Disaster - Skrach Uglogwee", parseQuestWidget("You have freed Skrach 'Bone Crusher' Uglogwee!"));
+        assertEquals("Recipe for Disaster", parseQuestWidget("You have completed Recipe for Disaster!"));
+
+        assertEquals("Doric's Quest", parseQuestWidget("You have completed Doric's Quest!"));
+        assertEquals("Heroes Quest", parseQuestWidget("You have completed the Heroes' Quest!"));
+
+        assertEquals("Dragon Slayer II", parseQuestWidget("You have completed Dragon Slayer II!"));
+        assertEquals("Rag and Bone Man II", parseQuestWidget("You have completed Rag and Bone Man II!"));
+        assertEquals("Fairytale II - Cure a Queen", parseQuestWidget("You have completed Fairytale II - Cure a Queen!"));
+
+        assertNull(parseQuestWidget("You have... kind of... completed Hazeel Cult!"));
+        assertEquals("Hazeel Cult", parseQuestWidget("You have completed Hazeel Cult!"));
     }
 
     private static WidgetLoaded event(int id) {